Create Stat superclass
